requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
aiohttp>=3.8.5
aiofiles>=23.1.0
mmh3>=4.0.0
pyyaml>=6.0